greater overhaul around unicode support our printers internally convert the string to utf32 added unifont.ttf for testing unicode everything else has just strings making font.cpp consistent with what imgui loads - we'll have to remove duplication here settings gui can tweak experimentally the font size imgui can throw error on init make main reload the gui font if the input has changed

+++ src/augs/utf32_point.h
@@ -1,3 +1,3 @@
 #pragma once
 
-using utf32_point = unsigned long;
+using utf32_point = unsigned int;
